# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type MemoryExtractedData = { [key: string]: unknown };

export interface Memory {
  id: string;
  originalText: string;
  currentText: string;
  category: string;
  extractedData: MemoryExtractedData;
  createdAt: string;
  updatedAt: string;
}

export interface MemoryInput {
  /** @minLength 1 */
  text: string;
  category?: string;
}

export interface MemoryUpdate {
  /** @minLength 1 */
  currentText: string;
}

export interface AssistantQuery {
  /** @minLength 1 */
  message: string;
}

export type AssistantResponseAction = typeof AssistantResponseAction[keyof typeof AssistantResponseAction];


export const AssistantResponseAction = {
  saved: 'saved',
  repeated: 'repeated',
  corrected: 'corrected',
  answered: 'answered',
  no_match: 'no_match',
} as const;

export interface AssistantResponse {
  reply: string;
  action: AssistantResponseAction;
  memory: Memory | null;
  memories?: Memory[];
}

export interface MemorySummary {
  total: number;
  today: number;
  latest: Memory | null;
}

export interface Error {
  error: string;
}

/**
 * Memory not found
 */
export type NotFoundResponse = Error;

export type ListMemoriesParams = {
search?: string;
/**
 * @minimum 1
 * @maximum 100
 */
limit?: number;
};