# lib/api-zod/src/generated/types/assistantResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AssistantResponseAction } from './assistantResponseAction';
import type { Memory } from './memory';

export interface AssistantResponse {
  reply: string;
  action: AssistantResponseAction;
  memory: Memory | null;
  memories?: Memory[];
}