# lib/api-zod/src/generated/types/memory.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { MemoryExtractedData } from './memoryExtractedData';

export interface Memory {
  id: string;
  originalText: string;
  currentText: string;
  category: string;
  extractedData: MemoryExtractedData;
  createdAt: Date;
  updatedAt: Date;
}