# lib/api-zod/src/generated/types/memoryInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface MemoryInput {
  /** @minLength 1 */
  text: string;
  category?: string;
}