# lib/api-zod/src/generated/types/memorySummary.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Memory } from './memory';

export interface MemorySummary {
  total: number;
  today: number;
  latest: Memory | null;
}